refactoring for - WinMain() - Common.h => v8cmn.h
