+++ NaMacro/KeyboardModule.cpp
@@ -43,7 +43,7 @@ Local<Object> NaKeyboardModule::GetKeyboardObject(Isolate *isolate)
 {
 	// HandleScope 안에서 호출
 
-	Local<Object> system_obj = GetSystemObject(isolate);
+	Local<Object> system_obj = v8cmn::GetSystemObject(isolate);
 	Local<String> keyboard_name = String::NewFromUtf8(isolate, "keyboard", NewStringType::kNormal).ToLocalChecked();
 	Local<Value> keyboard_value = system_obj->Get(keyboard_name);
 	if (!keyboard_value.IsEmpty() && keyboard_value->IsUndefined())
