+++ NaMacro/ModuleBase.h
@@ -1,7 +1,7 @@
 #pragma once
 
-#include "Common.h"
-#include "v8.h"
+#include "v8cmn.h"
+#include <v8.h>
 
 class ModuleBase
 {
