+++ NaMacro/MouseModule.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <v8.h>
-#include "Common.h"
+#include "v8cmn.h"
 #include "ModuleBase.h"
 
 class NaMouseModule : public ModuleBase
