+++ NaMacro/NaFile.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "Common.h"
+#include "v8cmn.h"
 #include "NaObject.h"
 
 #include <Windows.h>
