+++ NaMacro/NaMacro.vcxproj
@@ -142,7 +142,7 @@ move $(SolutionDir)..\NaMacroJS_Release\Temp\*.zip $(SolutionDir)..\NaMacroJS_Re
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClCompile Include="BasicModule.cpp" />
-    <ClCompile Include="Common.cpp" />
+    <ClCompile Include="v8cmn.cpp" />
     <ClCompile Include="ExtModule.cpp" />
     <ClCompile Include="KeyboardModule.cpp" />
     <ClCompile Include="MouseModule.cpp" />
@@ -163,7 +163,7 @@ move $(SolutionDir)..\NaMacroJS_Release\Temp\*.zip $(SolutionDir)..\NaMacroJS_Re
   <ItemGroup>
     <ClInclude Include="..\NaLib\NaHttp.h" />
     <ClInclude Include="BasicModule.h" />
-    <ClInclude Include="Common.h" />
+    <ClInclude Include="v8cmn.h" />
     <ClInclude Include="ExtModule.h" />
     <ClInclude Include="KeyboardModule.h" />
     <ClInclude Include="MouseModule.h" />
