+++ NaMacro/NaWindow.h
@@ -1,10 +1,10 @@
 #pragma once
 
-#include "Common.h"
+#include "v8cmn.h"
 #include "NaObject.h"
 
 #include <Windows.h>
-#include "v8.h"
+#include <v8.h>
 
 #include <list>
 #include <map>
