+++ NaMacro/v8cmn.h
@@ -4,22 +4,51 @@
 #include <NaLib/NaString.h>
 #include <NaLib/NaUrl.h>
 
-#define MOUSECLICK_SLEEP	10
-#define NA_DEBUGOUT_TEMPBUFFER_SIZE	65535
+namespace v8cmn {
+	using namespace v8;
 
-extern bool g_bReportExceptions;
-extern bool g_bExit;
+	extern bool g_bReportExceptions;
+	extern bool g_bExit;
 
-using namespace v8;
+	Platform* Initialize();
+	Isolate* CreateNewIsolate();
+	void TearDown(Platform* platform);
 
-// v8 Utility
-NaString ReportException(Isolate *isolate, TryCatch* handler);
+	void CreateDefaultModules(Isolate *isolate, Local<ObjectTemplate> &global_template);
+	void InitModules(Isolate *isolate, Local<ObjectTemplate> &global_template);
+	void ReleaseModules();
 
-// Utility Functions
-Local<String> ReadFile(Isolate *isolate, const char* name);
-Local<Object> GetSystemObject(Isolate *isolate);
+	Local<String> ReadFile(Isolate *isolate, const char* name);
+	Local<String> ReadScript(Isolate* isolate, const std::string& scriptFilePath);
+	Local<Script> Compile(Isolate* isolate, Local<Context>& context,
+							const std::string& scriptPath,
+							Local<String>& scriptSource);
+	bool RunScript(Isolate* isolate, Local<Context>& context, Local<Script>& script);
+	bool RunMainFunc(Isolate* isolate);
+	void MessageLoopTillExit(Isolate* isolate);
 
-// Macro Defines
+	Local<Object> GetSystemObject(Isolate *isolate);
+	NaString ReportException(Isolate *isolate, TryCatch* handler);
+
+	// Buffer Allocator
+	class ArrayBufferAllocator : public ArrayBuffer::Allocator {
+	public:
+		virtual void* Allocate(size_t length) {
+			void* data = AllocateUninitialized(length);
+			return data == NULL ? data : memset(data, 0, length);
+		}
+		virtual void* AllocateUninitialized(size_t length) {
+			return malloc(length);
+		}
+		virtual void Free(void* data, size_t) {
+			free(data);
+		}
+	};
+}
+
+//
+// Macro Defines 
+//
 #define V8_FUNCTION_ARGS	const FunctionCallbackInfo<Value>& args
 
 #define V8_GETTER_ARGS		Local<String> name, const PropertyCallbackInfo<Value>& info
